Start generalizing code to support more datasets
The initial release of the CDS Plugin was, in part, hard-coded for a
single ERA5 dataset. This commit begins to generalize that code,
making it easier to support additional datasets. Most noticeably, the
ERA5 variable definitions are no longer contained in constants.py, but
have been converted to JSON and placed in a datasets subdirectory
organized on a "one JSON file per dataset" basis, from where they are
read at runtime.