Adjust ERA5 bounding box to match xcube semantics
xcube expects that each data point is the centre of a cell. ERA5 data
are in fact point values. This difference becomes important when
calculating a bounding box: the bounding box of a grid of cell centres
is half a cell-width beyond the grid's edge, but the bounding box of a
grid of zero-dimensional points is precisely *at* the grid's edge.
This means that, for a global bounding box specification of [-180,
-90, 180, 90], ERA5 will return a grid going right up to those extrema
(except for +180, since this is the same as -180). xcube will
helpfully extend those grid points by half a cell-width to (in the
case of a 0.25 degree ggrid) -180.125, -90.125, and +90.125, which is
of course invalid.

To correct this, the CDS plugin now shrinks the size specification by
half a cell-width in every direction before making the request, resulting
in a final bounding box which matches the one originally requested.

For more on ERA5 grid points, see
https://confluence.ecmwf.int/display/CKB/ERA5%3A+What+is+the+spatial+reference#ERA5:Whatisthespatialreference-Visualisationofregularlat/londata .