Move param transform & file read to DatasetHandler
This commit continues the generalization of dataset handling: the
open_data method in CDSDataStore now calls through to a read_file
method in the handler for the particular requested dataset.
Communication with the CDS API and normalization are still carried out
in the open_data method. The implementation of read_file in the ERA5
dataset handler just makes a single call to xarray.open_dataset,
but things will be more complex for datasets which return an archive
of multiple data files.

Similarly, the transform_params method, which generates CDS API
parameters from the xcube opener parameters, has been moved into
DatasetHandler, since the request format varies widely among different
datasets.