Partial implementation of soil moisture dataset
This commit provides functional, but very restricted, support for the
satellite soil moisture dataset. As with the ERA5 datasets, it is
supported as multiple virtual sub-datasets to deal with the fact that
the data format is dependent on some of the request parameters.
get_open_data_params_schema does not yet take account of this fact.
Most significantly, read_data only makes use of one of the NetCDF
files in the returned archive, so only one time-point can be retrieved
per call.