More explicit parameters in xr.open_dataset calls
Specify the "engine" and "decode_cf" parameters explicitly in calls to
open_dataset and open_mfdataset. This doesn't change current behaviour,
but makes the calls more robust against changes in API or environment,
and in the case of "engine" provides a more meaningful error message when
a specified input file is missing.