Update unit tests and mock results
Update some tests, both to match the new ERA5 behaviour requested
in Issue 99, and to accommodate changes in the behaviour of the
live CDS API server. The new mock results were produced by running
the test suite with mock._BEHAVIOUR set to SAVE_RESULTS. Note that
not all API updates are covered yet -- in particular,
test_open_data_null_variables_list fails with the current live API
and has therefore not had its mock results updated. This is probably
indicative of a bigger problem, for which I will open a new issue.